+++ config/check_vaa_err.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+# filepath: check_vaa_err.sh
+
+grep -Hni "\[Erro" /data/VAH/vae0/product/log/partition/*.txt