+++ config/check_vah_err.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# filepath: check_vah_err.sh
+ 
+grep -Hni "Error" /data/VAH/vae0/log/events/*.qvt
+