+++ config/debug_vae0.sh
@@ -1,0 +1,7 @@
+rm /data/smf/temp/psdpartition/vae0.bin
+ 
+rm /data/smf/data/MultiConfigurations/Application/vae0.xml
+ 
+rm /data/smf/data/MultiConfigurations/PsdPartition/vae0.xml
+ 
+rm /data/smf/data/MultiConfigurations/SmfStatsCollector/vae0.xml